Address conversion warnings in essimporter

+++ apps/essimporter/convertacdt.cpp
@@ -48,8 +48,8 @@ namespace ESSImport
         for (int i = 0; i < ESM::Skill::Length; ++i)
         {
             npcStats.mSkills[i].mMod = 0.f;
-            npcStats.mSkills[i].mCurrent = actorData.mSkills[i][1];
-            npcStats.mSkills[i].mBase = actorData.mSkills[i][0];
+            npcStats.mSkills[i].mCurrent = static_cast<float>(actorData.mSkills[i][1]);
+            npcStats.mSkills[i].mBase = static_cast<float>(actorData.mSkills[i][0]);
         }
 
         npcStats.mTimeToStartDrowning = actorData.mACDT.mBreathMeter;
